feat(app-shell): bootstrap client app to work with shared dependencies
