+++ apps/AppShell/src/client/index.tsx
@@ -1,36 +1,8 @@
-// import ReactDOM from 'react-dom'
-import {Provider as ReduxProvider} from 'react-redux'
+/**
+ * This import structure is required so Webpack splits chunks correctly, so shared
+ * modules are loaded when running the micro-frontend on its own.
+ * See {@link https://webpack.js.org/concepts/module-federation/#uncaught-error-shared-module-is-not-available-for-eager-consumption}
+ */
 
-import {Logger} from '@batman/core-logger'
-import {configureStore} from '@reduxjs/toolkit'
-import {hydrateRoot} from 'react-dom/client'
-
-import counterSlice from '../ducks/counter'
-import reportWebVitals from '../reportWebVitals'
-import {spliceAllToHead} from '../utils/spliceAllToHead'
-import App from './client_app'
-
-const links = document.body.getElementsByTagName('link')
-const scripts = document.body.getElementsByTagName('script')
-
-spliceAllToHead(links)
-spliceAllToHead(scripts)
-
-const root = document.getElementById('federated_modules_root_id')
-
-// TODO: type this properly
-// TODO: generate from arbitrary reducers
-const store = configureStore({reducer: {counter: counterSlice}, preloadedState: (window as any).initialState})
-delete (window as any).initialState
-
-hydrateRoot(
-  root!, // eslint-disable-line @typescript-eslint/no-non-null-assertion
-  <ReduxProvider store={store}>
-    <App />
-  </ReduxProvider>,
-)
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals(Logger.info)
+import('./bootstrap')
+export default {}
